refactor(webapp): move PROD_ENVIRONMENT_NAME into store/index.ts
