--- packages/webapp/src/utils/constants.ts
@@ -1,1 +1,0 @@
-export const PROD_ENVIRONMENT_NAME = 'prod';
